requirements.txt
requests
BeatifulSoup4
discord.py
